test: Move `test_preboot_load_snap_disallowed_after_boot_resources`
Move this test from being a mock-based test in rpc_interface.rs to
integration_test.rs, where no mocking is needed.

While doing this, drop the MMDS case, as it's impossible to even make
the MMDS request outside of the mock context, because it requires a
network interface to be set up first, and we already test that loading
snapshots after setting up network interfaces is impossible.
