+++ src/vmm/src/rpc_interface.rs
@@ -1646,87 +1646,4 @@ mod tests {
             check_unsupported,
         );
     }
-
-    fn verify_load_snap_disallowed_after_boot_resources(res: VmmAction, res_name: &str) {
-        let mut vm_resources = MockVmRes::default();
-        let mut evmgr = EventManager::new().unwrap();
-        let seccomp_filters = BpfThreadMap::new();
-        let mut preboot = default_preboot(&mut vm_resources, &mut evmgr, &seccomp_filters);
-
-        preboot.handle_preboot_request(res).unwrap();
-
-        // Load snapshot should no longer be allowed.
-        let req = VmmAction::LoadSnapshot(LoadSnapshotParams {
-            snapshot_path: PathBuf::new(),
-            mem_backend: MemBackendConfig {
-                backend_type: MemBackendType::File,
-                backend_path: PathBuf::new(),
-            },
-            enable_diff_snapshots: false,
-            resume_vm: false,
-        });
-        let err = preboot.handle_preboot_request(req);
-        assert!(
-            matches!(
-                err.unwrap_err(),
-                VmmActionError::LoadSnapshot(LoadSnapshotError::LoadSnapshotNotAllowed)
-            ),
-            "LoadSnapshot should be disallowed after {}",
-            res_name
-        );
-    }
-
-    #[test]
-    fn test_preboot_load_snap_disallowed_after_boot_resources() {
-        // Verify LoadSnapshot not allowed after configuring various boot-specific resources.
-        let req = VmmAction::ConfigureBootSource(BootSourceConfig::default());
-        verify_load_snap_disallowed_after_boot_resources(req, "ConfigureBootSource");
-
-        let config = BlockDeviceConfig {
-            drive_id: String::new(),
-            partuuid: None,
-            is_root_device: false,
-            cache_type: CacheType::Unsafe,
-
-            is_read_only: Some(false),
-            path_on_host: Some(String::new()),
-            rate_limiter: None,
-            file_engine_type: None,
-
-            socket: None,
-        };
-
-        let req = VmmAction::InsertBlockDevice(config);
-        verify_load_snap_disallowed_after_boot_resources(req, "InsertBlockDevice");
-
-        let req = VmmAction::InsertNetworkDevice(NetworkInterfaceConfig {
-            iface_id: String::new(),
-            host_dev_name: String::new(),
-            guest_mac: None,
-            rx_rate_limiter: None,
-            tx_rate_limiter: None,
-        });
-        verify_load_snap_disallowed_after_boot_resources(req, "InsertNetworkDevice");
-
-        let req = VmmAction::SetBalloonDevice(BalloonDeviceConfig::default());
-        verify_load_snap_disallowed_after_boot_resources(req, "SetBalloonDevice");
-
-        let req = VmmAction::SetVsockDevice(VsockDeviceConfig {
-            vsock_id: Some(String::new()),
-            guest_cid: 0,
-            uds_path: String::new(),
-        });
-        verify_load_snap_disallowed_after_boot_resources(req, "SetVsockDevice");
-
-        let req =
-            VmmAction::UpdateVmConfiguration(MachineConfigUpdate::from(MachineConfig::default()));
-        verify_load_snap_disallowed_after_boot_resources(req, "SetVmConfiguration");
-
-        let req = VmmAction::SetMmdsConfiguration(MmdsConfig {
-            ipv4_address: None,
-            version: MmdsVersion::default(),
-            network_interfaces: Vec::new(),
-        });
-        verify_load_snap_disallowed_after_boot_resources(req, "SetMmdsConfiguration");
-    }
 }
