+++ src/vmm/tests/integration_tests.rs
@@ -7,19 +7,28 @@ use std::time::Duration;
 
 use utils::tempfile::TempFile;
 use vmm::builder::build_and_boot_microvm;
+use vmm::devices::virtio::block::CacheType;
 use vmm::persist::{self, snapshot_state_sanity_check, MicrovmState, MicrovmStateError, VmInfo};
 use vmm::resources::VmResources;
-use vmm::rpc_interface::{PrebootApiController, RuntimeApiController, VmmAction};
+use vmm::rpc_interface::{
+    LoadSnapshotError, PrebootApiController, RuntimeApiController, VmmAction, VmmActionError,
+};
 use vmm::seccomp_filters::get_empty_filters;
 use vmm::snapshot::Snapshot;
 use vmm::utilities::mock_resources::{MockVmResources, NOISY_KERNEL_IMAGE};
 #[cfg(target_arch = "x86_64")]
 use vmm::utilities::test_utils::dirty_tracking_vmm;
 use vmm::utilities::test_utils::{create_vmm, default_vmm, default_vmm_no_boot};
+use vmm::vmm_config::balloon::BalloonDeviceConfig;
+use vmm::vmm_config::boot_source::BootSourceConfig;
+use vmm::vmm_config::drive::BlockDeviceConfig;
 use vmm::vmm_config::instance_info::{InstanceInfo, VmState};
+use vmm::vmm_config::machine_config::{MachineConfig, MachineConfigUpdate};
+use vmm::vmm_config::net::NetworkInterfaceConfig;
 use vmm::vmm_config::snapshot::{
     CreateSnapshotParams, LoadSnapshotParams, MemBackendConfig, MemBackendType, SnapshotType,
 };
+use vmm::vmm_config::vsock::VsockDeviceConfig;
 use vmm::{DumpCpuConfigError, EventManager, FcExitCode};
 
 #[test]
@@ -303,4 +312,92 @@ fn get_microvm_state_from_snapshot() -> MicrovmState {
     snapshot_file.as_file().seek(SeekFrom::Start(0)).unwrap();
     let (state, _) = Snapshot::load(&mut snapshot_file.as_file(), snapshot_len).unwrap();
     state
+}
+
+fn verify_load_snap_disallowed_after_boot_resources(res: VmmAction, res_name: &str) {
+    let (snapshot_file, memory_file) = verify_create_snapshot(false);
+
+    let mut event_manager = EventManager::new().unwrap();
+    let empty_seccomp_filters = get_empty_filters();
+    let mut vm_resources = VmResources::default();
+
+    let mut preboot_api_controller = PrebootApiController::new(
+        &empty_seccomp_filters,
+        InstanceInfo::default(),
+        &mut vm_resources,
+        &mut event_manager,
+    );
+
+    preboot_api_controller.handle_preboot_request(res).unwrap();
+
+    // Load snapshot should no longer be allowed.
+    let req = VmmAction::LoadSnapshot(LoadSnapshotParams {
+        snapshot_path: snapshot_file.as_path().to_path_buf(),
+        mem_backend: MemBackendConfig {
+            backend_path: memory_file.as_path().to_path_buf(),
+            backend_type: MemBackendType::File,
+        },
+        enable_diff_snapshots: false,
+        resume_vm: false,
+    });
+    let err = preboot_api_controller.handle_preboot_request(req);
+    assert!(
+        matches!(
+            err.unwrap_err(),
+            VmmActionError::LoadSnapshot(LoadSnapshotError::LoadSnapshotNotAllowed)
+        ),
+        "LoadSnapshot should be disallowed after {}",
+        res_name
+    );
+}
+
+#[test]
+fn test_preboot_load_snap_disallowed_after_boot_resources() {
+    let tmp_file = TempFile::new().unwrap();
+    let tmp_file = tmp_file.as_path().to_str().unwrap().to_string();
+    // Verify LoadSnapshot not allowed after configuring various boot-specific resources.
+    let req = VmmAction::ConfigureBootSource(BootSourceConfig {
+        kernel_image_path: tmp_file.clone(),
+        ..Default::default()
+    });
+    verify_load_snap_disallowed_after_boot_resources(req, "ConfigureBootSource");
+
+    let config = BlockDeviceConfig {
+        drive_id: String::new(),
+        partuuid: None,
+        is_root_device: false,
+        cache_type: CacheType::Unsafe,
+
+        is_read_only: Some(false),
+        path_on_host: Some(tmp_file),
+        rate_limiter: None,
+        file_engine_type: None,
+
+        socket: None,
+    };
+
+    let req = VmmAction::InsertBlockDevice(config);
+    verify_load_snap_disallowed_after_boot_resources(req, "InsertBlockDevice");
+
+    let req = VmmAction::InsertNetworkDevice(NetworkInterfaceConfig {
+        iface_id: String::new(),
+        host_dev_name: String::new(),
+        guest_mac: None,
+        rx_rate_limiter: None,
+        tx_rate_limiter: None,
+    });
+    verify_load_snap_disallowed_after_boot_resources(req, "InsertNetworkDevice");
+
+    let req = VmmAction::SetBalloonDevice(BalloonDeviceConfig::default());
+    verify_load_snap_disallowed_after_boot_resources(req, "SetBalloonDevice");
+
+    let req = VmmAction::SetVsockDevice(VsockDeviceConfig {
+        vsock_id: Some(String::new()),
+        guest_cid: 0,
+        uds_path: String::new(),
+    });
+    verify_load_snap_disallowed_after_boot_resources(req, "SetVsockDevice");
+
+    let req = VmmAction::UpdateVmConfiguration(MachineConfigUpdate::from(MachineConfig::default()));
+    verify_load_snap_disallowed_after_boot_resources(req, "SetVmConfiguration");
 }
